+++ tests/Support/config/.merge-plan.php
@@ -10,9 +10,9 @@ return [
                 'params.php',
             ],
         ],
-        'web' => [
+        'di-web' => [
             '/' => [
-                'web.php',
+                'di-web.php',
             ],
         ],
         'events-web' => [
@@ -20,14 +20,14 @@ return [
                 'events-web.php',
             ],
         ],
-        'providers-web' => [
+        'di-providers-web' => [
             '/' => [
-                'providers-web.php',
+                'di-providers-web.php',
             ],
         ],
-        'delegates-web' => [
+        'di-delegates-web' => [
             '/' => [
-                'delegates-web.php',
+                'di-delegates-web.php',
             ],
         ],
         'bootstrap-web' => [